[#24] Moved Sample data from Application file to data directory
